In production.
Check the configuration setting at the top of the code. The pins D5 and D6 and DHT11 are not those used in the Luftdaten version so you may need to change the code before compiling. Cycle and warm up times can be changed at reboot. You may also want to change User and password values.